LocationId displays on fish, hasFake displays, birthdays as well
