+++ Pocketpedia/Repositories/FishRepository.cs
@@ -13,11 +13,14 @@ namespace Pocketpedia.Repositories
 {
     public class FishRepository : BaseRepository, IFishRepository
     {
-        public FishRepository(IConfiguration configuration, ILocationRepository locationRepository) : base(configuration) { }
+        public FishRepository(IConfiguration configuration, ILocationRepository locationRepository) : base(configuration)
+        {
+            this.locationRepository = locationRepository;
+        }
 
-        private static readonly HttpClient client = new HttpClient();
+    private static readonly HttpClient client = new HttpClient();
 
-       // private readonly ILocationRepository locationRepository;
+        private readonly ILocationRepository locationRepository;
 
         public async Task<List<Fish>> FishesFromApi()
         {
@@ -28,14 +31,15 @@ namespace Pocketpedia.Repositories
             var response = await client.GetStreamAsync($"http://acnhapi.com/v1/fish");
             var apiFishes = await JsonSerializer.DeserializeAsync<Dictionary<string, ApiFish>>(response);
 
-            //var locations = locationRepository.GetLocations();
+            var locations = locationRepository.GetLocations();
+
 
             var desiredResponse = apiFishes.Values.Select(apiFish => new Fish()
             {
                 AcnhApiId = apiFish.id,
                 Name = apiFish.filename,
-               // LocationId = locations.FirstOrDefault(location => apiFish.availability.location == location.Name).Id,
-                ImageUrl = apiFish.image_uri
+                LocationId = locations.FirstOrDefault(location => apiFish.availability.location == location.Name).Id,
+                ImageUrl = apiFish.icon_uri
             }).ToList();
 
             return desiredResponse;
