+++ Pocketpedia/client/src/components/Bugs/BugCard.js
@@ -3,13 +3,15 @@ import { Card, CardBody } from "reactstrap";
 
 const BugCard = ({ bug }) => {
 
+   const locationId = bug.locationId.toString();
+
     return (
         <Card className="m-2 p-2 w-50 mx-auto">
             <CardBody className="m-3">
 
                 <img src={bug.imageUrl} alt={bug.name} />
                 <p><b>Title: </b>{bug.name}</p>
-                <p><b>Location: </b>{bug.locationId}</p>
+                <p><b>Location: </b>{locationId}</p>
 
             </CardBody>
         </Card>
