+++ Pocketpedia/client/src/components/Villagers/VillagerCard.js
@@ -3,6 +3,7 @@ import { Card, CardBody } from "reactstrap";
 
 const VillagerCard = ({ villagers }) => {
 
+
     return (
         <Card className="m-2 p-2 w-50 mx-auto">
             <CardBody className="m-3">
