+++ Pocketpedia/client/src/modules/locationManager.js
@@ -1,0 +1,3 @@
+import { getToken } from "./authManager";
+
+const baseUrl = '/api/location';
